+++ packages/example/pages/_app.tsx
@@ -1,12 +1,14 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable unused-imports/no-unused-imports */
 import "bootstrap/dist/css/bootstrap.min.css";
+import "@leapwallet/cosmos-social-login-capsule-provider-ui/styles.css";
 import "../style/global.css";
 import "@interchain-ui/react/styles";
 
 import { Chain } from "@chain-registry/types";
 import { Decimal } from "@cosmjs/math";
 import { GasPrice } from "@cosmjs/stargate";
+import { wallets as cdcwalletWallets } from "@cosmos-kit/cdcwallet";
 import { wallets as coin98Wallets } from "@cosmos-kit/coin98";
 import { ChainName } from "@cosmos-kit/core";
 import { MainWalletBase } from "@cosmos-kit/core";
@@ -14,25 +16,25 @@ import { wallets as cosmostationWallets } from "@cosmos-kit/cosmostation";
 import { wallets as exodusWallets } from "@cosmos-kit/exodus";
 import { wallets as finWallets } from "@cosmos-kit/fin";
 import { wallets as frontierWallets } from "@cosmos-kit/frontier";
+import { wallets as galaxyStationWallets } from "@cosmos-kit/galaxy-station";
 import { wallets as keplrWallets } from "@cosmos-kit/keplr";
-import { wallets as owalletWallets } from "@cosmos-kit/owallet";
 import { wallets as leapWallets } from "@cosmos-kit/leap";
 import { wallets as snapWallet } from "@cosmos-kit/leap-metamask-cosmos-snap";
 import { wallets as ledgerWallets } from "@cosmos-kit/ledger";
 import { wallets as ninjiWallets } from "@cosmos-kit/ninji";
 import { wallets as omniWallets } from "@cosmos-kit/omni";
+import { wallets as owalletWallets } from "@cosmos-kit/owallet";
 // Show how to custom modal views
 import { ChainProvider, defaultModalViews } from "@cosmos-kit/react";
 // import { ChainProvider } from "@cosmos-kit/react";
 import { wallets as shellWallets } from "@cosmos-kit/shell";
 import { wallets as stationWallets } from "@cosmos-kit/station";
-import { wallets as galaxyStationWallets } from "@cosmos-kit/galaxy-station";
 import { wallets as tailwindWallet } from "@cosmos-kit/tailwind";
 import { wallets as trustWallets } from "@cosmos-kit/trust";
 import { wallets as vectisWallets } from "@cosmos-kit/vectis";
 // import { makeWeb3AuthWallets } from "@cosmos-kit/web3auth";
 import { wallets as xdefiWallets } from "@cosmos-kit/xdefi";
-import { wallets as cdcwalletWallets } from "@cosmos-kit/cdcwallet";
+import { useTheme } from "@interchain-ui/react";
 import { assets, chains } from "chain-registry";
 import type { AppProps } from "next/app";
 import dynamic from "next/dynamic";
@@ -40,7 +42,7 @@ import React, { useEffect, useMemo, useState } from "react";
 
 // import { CustomConnectedView } from "../components/custom-connected";
 import { RootLayout } from "../components/layout";
-// import { useTheme } from "@interchain-ui/react";
+
 
 function MyApp({ Component, pageProps }: AppProps) {
   const defaultWallets: MainWalletBase[] = [...keplrWallets, ...leapWallets];
@@ -88,7 +90,7 @@ function MyApp({ Component, pageProps }: AppProps) {
       <ChainProvider
         // chains={chains}
         // assetLists={[...assets]}
-        chains={["cosmoshub", "secret"]}
+        chains={["cosmoshub", "secretnetwork"]}
         assetLists={[]}
         wallets={[
           // ...wallets,
@@ -213,8 +215,10 @@ export function CustomCapsuleModalViewX() {
     );
   }, []);
 
+  const { theme } = useTheme();
+
   return (
-    <>
+    <div className={`leap-ui z-[9999] fixed ${theme === "dark" ? "dark" : ""}`}>
       <LeapSocialLogin
         showCapsuleModal={showCapsuleModal}
         setShowCapsuleModal={setShowCapsuleModal}
@@ -230,6 +234,6 @@ export function CustomCapsuleModalViewX() {
         appName="CosmosKit"
         oAuthMethods={oAuthMethods}
       />
-    </>
+    </div>
   );
 }