+++ packages/react-lite/src/provider.tsx
@@ -36,9 +36,14 @@ export function ChainProvider({
   sessionOptions,
   logLevel = 'WARN',
   allowedIframeParentOrigins = [
+    'http://localhost:*',
+    'https://localhost:*',
     'https://app.osmosis.zone',
     'https://daodao.zone',
     'https://dao.daodao.zone',
+    'https://my.abstract.money',
+    'https://apps.abstract.money',
+    'https://console.abstract.money'
   ],
   children,
 }: {