+++ wallets/imtoken/src/index.ts
@@ -1,0 +1,3 @@
+import { wallets as ext } from '@cosmos-kit/imtoken-extension';
+
+export const wallets = [...ext];