+++ wallets/leap-capsule-social-login/src/chain-wallet.ts
@@ -1,7 +1,7 @@
-import { ChainRecord, ChainWalletBase,Wallet } from '@cosmos-kit/core'
+import { ChainRecord, ChainWalletBase, Wallet } from '@cosmos-kit/core';
 
 export class ChainCosmosSocial extends ChainWalletBase {
   constructor(walletInfo: Wallet, chainInfo: ChainRecord) {
-    super(walletInfo, chainInfo)
+    super(walletInfo, chainInfo);
   }
 }