+++ packages/core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.13.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.13.0...@cosmos-kit/core@2.13.1) (2024-06-28)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
+
+
+
+
+# [2.13.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.12.0...@cosmos-kit/core@2.13.0) (2024-06-18)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
+
+
+
+
 # [2.12.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.11.0...@cosmos-kit/core@2.12.0) (2024-05-28)
 
 **Note:** Version bump only for package @cosmos-kit/core